+++ src/plugins/intel_cpu/src/nodes/interpolate.cpp
@@ -14,16 +14,6 @@
 #include "ie_parallel.hpp"
 #include <algorithm>
 
-#include <cpu/x64/jit_generator.hpp>
-#include <cpu/x64/jit_uni_eltwise.hpp>
-#include <cpu/x64/injectors/jit_uni_depthwise_injector.hpp>
-#include <cpu/x64/injectors/jit_uni_quantization_injector.hpp>
-#include <cpu/x64/injectors/jit_uni_eltwise_injector.hpp>
-#include "common/cpu_memcpy.h"
-#include "utils/bfloat16.hpp"
-#include "emitters/x64/jit_bf16_emitters.hpp"
-#include "emitters/x64/jit_load_store_emitters.hpp"
-
 #include <ngraph/opsets/opset1.hpp>
 #include <ngraph/opsets/opset4.hpp>
 #include <utils/shape_inference/static_shape.hpp>
@@ -32,1345 +22,19 @@
 #include "utils/cpu_utils.hpp"
 #include <utils/shape_inference/shape_inference_ngraph.hpp>
 
-using namespace dnnl;
-using namespace InferenceEngine;
-using namespace dnnl::impl;
-using namespace dnnl::impl::cpu;
-using namespace dnnl::impl::cpu::x64;
-using namespace dnnl::impl::utils;
-using namespace Xbyak;
-
-
-#define GET_OFF(field) offsetof(jit_interpolate_call_args, field)
+#include "common/cpu_memcpy.h"
+#include "utils/bfloat16.hpp"
+#include "emitters/x64/jit_load_store_emitters.hpp"
 
 namespace ov {
 namespace intel_cpu {
 namespace node {
-
-template <cpu_isa_t isa>
-struct jit_uni_interpolate_kernel_f32 : public jit_uni_interpolate_kernel, public jit_generator {
-    DECLARE_CPU_JIT_AUX_FUNCTIONS(jit_uni_interpolate_kernel_f32)
-
-    explicit jit_uni_interpolate_kernel_f32(jit_interpolate_config_params jcp, const dnnl_primitive_attr &attr)
-    : jit_uni_interpolate_kernel(jcp, attr), jit_generator(jit_name()) {}
-
-    void create_ker() override {
-        jit_generator::create_kernel();
-        ker_ = (decltype(ker_))jit_ker();
-    }
-
-    void generate() override {
-        // dummy second reg_tmp_64 as no fill needed
-        load_pool_gpr_idxs = {static_cast<size_t>(reg_tmp_64.getIdx()), static_cast<size_t>(reg_tmp_64.getIdx())};
-        store_pool_gpr_idxs = {static_cast<size_t>(reg_tmp_64.getIdx())};
-        store_pool_vec_idxs = {static_cast<size_t>(vmm_zero.getIdx())};
-
-        const auto &p = attr_.post_ops_;
-        for (int i = 0; i < p.len(); i++) {
-            auto &post_op = p.entry_[i];
-            if (post_op.is_eltwise()) {
-                eltwise_injectors.push_back(std::make_shared<jit_uni_eltwise_injector_f32<isa>>(
-                        this,
-                        post_op.eltwise.alg,
-                        post_op.eltwise.alpha,
-                        post_op.eltwise.beta,
-                        1.f));
-            } else if (post_op.is_depthwise()) {
-                depthwise_injectors.push_back(std::make_shared<jit_uni_depthwise_injector_f32<isa>>(
-                        this,
-                        post_op));
-            } else if (post_op.is_quantization()) {
-                quantization_injectors.push_back(std::make_shared<jit_uni_quantization_injector_f32<isa>>(
-                        this, post_op, vmm_d_weights, vmm_d_bias, reg_d_weights, reg_d_bias));
-            }
-        }
-
-        this->preamble();
-
-        if (attr_.post_ops_.len() != 0) {
-            mov(reg_post_ops_data, ptr[reg_params + GET_OFF(post_op_data)]);
-            mov(reg_oc_off, ptr[reg_params + GET_OFF(oc_off)]);
-        }
-        uni_vpxor(vmm_zero, vmm_zero, vmm_zero);
-
-        switch (jcp_.mode) {
-            case InterpolateMode::nearest: {
-                mov(reg_dst, ptr[reg_params + GET_OFF(dst)]);
-                mov(reg_src, ptr[reg_params + GET_OFF(src_ptr[0])]);
-                mov(reg_index, ptr[reg_params + GET_OFF(index)]);
-                mov(reg_work_amount, ptr[reg_params + GET_OFF(work_amount)]);
-
-                switch (jcp_.layout) {
-                    case InterpolateLayoutType::planar: {
-                        nn_planar();
-                        break;
-                    }
-                    case InterpolateLayoutType::block: {
-                        nn_blk();
-                        break;
-                    }
-                    case InterpolateLayoutType::by_channel: {
-                        nn_by_channel();
-                        break;
-                    }
-                    default:
-                        assert(!"unsupported memory layout for interpolate layer with nearest neighbor mode.");
-                }
-                break;
-            }
-            case InterpolateMode::linear_onnx: {
-                switch (jcp_.layout) {
-                    case InterpolateLayoutType::planar: {
-                        linear_onnx_planar();
-                        break;
-                    }
-                    case InterpolateLayoutType::block:
-                    case InterpolateLayoutType::by_channel: {
-                        linear_onnx_c_gathered();
-                        break;
-                    }
-                    default:
-                        assert(!"unsupported memory layout for interpolate layer with linear_onnx mode.");
-                }
-                break;
-            }
-            case InterpolateMode::cubic: {
-                switch (jcp_.layout) {
-                    case InterpolateLayoutType::planar: {
-                        cubic_planar();
-                        break;
-                    }
-                    case InterpolateLayoutType::block:
-                    case InterpolateLayoutType::by_channel: {
-                        cubic_c_gathered();
-                        break;
-                    }
-                    default:
-                        assert(!"unsupported memory layout for interpolate layer with cubic mode.");
-                }
-                break;
-            }
-            case InterpolateMode::linear: {
-                assert(!"unsupported mode for interpolate layer with JITTED implimentation.");
-                break;
-            }
-            default: {
-                assert(!"unsupported mode for interpolate layer.");
-            }
-        }
-
-        this->postamble();
-
-        emit_emitters_data();
-        for (auto& inj : eltwise_injectors)
-            inj->prepare_table();
-        if ((jcp_.mode == InterpolateMode::cubic) && (jcp_.layout == InterpolateLayoutType::planar)) {
-            prepare_cubic_planar_table();
-        }
-    }
-
-private:
-    using Vmm = typename conditional3<isa == cpu::x64::sse41, Xbyak::Xmm, isa == cpu::x64::avx2,
-            Xbyak::Ymm, Xbyak::Zmm>::type;
-
-    const int vlen = cpu_isa_traits<isa>::vlen;
-    const int vector_step = vlen / sizeof(float);
-    const int tail_step = jcp_.C % vector_step;
-    const int scalar_step = 1;
-
-    Xbyak::Reg64 reg_src = r8;
-    Xbyak::Reg64 reg_src_aux = r15;
-    Xbyak::Reg64 reg_src_aux1 = r11;
-    Xbyak::Reg64 reg_src_aux2 = r12;
-    Xbyak::Reg64 reg_dst = r9;
-    Xbyak::Reg64 reg_work_amount = r13;
-    Xbyak::Reg64 reg_index = r14;
-    Xbyak::Reg64 reg_params = abi_param1;
-
-    Reg8 reg_tmp_8 = r10b;
-    Reg32 reg_tmp_32 = r10d;
-    Reg64 reg_tmp_64 = r10;
-
-    Xbyak::Reg64 reg_oc_off = rax;
-    Xbyak::Reg64 reg_post_ops_data = rbx;
-    Xbyak::Reg64 reg_d_weights = reg_tmp_64;
-    Xbyak::Reg64 reg_d_bias = rcx;
-    Xbyak::Reg32 reg_index_offset = edx;
-
-    // for cubic planar
-    Xbyak::Reg64 reg_tbl_y = rsi;
-    Xbyak::Reg64 reg_tbl_x = rbp;
-    Xbyak::Reg64 reg_table = rdx;   // do not need reg_index_offset in this mode, so use rdx
-
-    Vmm vmm_val = Vmm(1);
-    Vmm vmm_index = Vmm(0);
-    Vmm vmm_zero = Vmm(2);
-    Vmm vmm_mask = Vmm(3);
-    Vmm vmm_d_weights = Vmm(4);
-    Vmm vmm_d_bias = Vmm(5);
-
-    // for linear
-    Vmm vmm_weightT = Vmm(15);
-    Vmm vmm_weightB = Vmm(14);
-    Vmm vmm_weightL = Vmm(13);
-    Vmm vmm_weightR = Vmm(12);
-    Vmm vmm_weightF = Vmm(6);
-    Vmm vmm_weightE = Vmm(7);
-    Vmm vmm_valTL = Vmm(11);
-    Vmm vmm_valTR = vmm_val;
-    Vmm vmm_valBL = Vmm(9);
-    Vmm vmm_valBR = Vmm(8);
-
-    // for cubic
-    Vmm vmm_src = Vmm(6);
-    Xmm xmm_src = Xmm(6);
-    Vmm vmm_dstX = Vmm(7);
-
-    Vmm vmm_weightX0 = vmm_weightT;
-    Vmm vmm_weightX1 = vmm_weightB;
-    Vmm vmm_weightX2 = vmm_weightL;
-    Vmm vmm_weightX3 = vmm_weightR;
-    Vmm vmm_weightY0 = vmm_valTL;
-    Vmm vmm_weightY1 = Vmm(10);  // vmm_valTR is vmm_val, need reserved
-    Vmm vmm_weightY2 = vmm_valBL;
-    Vmm vmm_weightY3 = vmm_valBR;
-    // cubic planar
-    Vmm vmm_one = vmm_index;
-    Vmm vmm_weightY = vmm_weightY0;
-    Vmm vmm_index_y_itr = vmm_weightY1;
-    Vmm vmm_index_x_itr = vmm_weightY2;
-    Vmm vmm_tbl_y = vmm_weightY3;
-    // temporally used. when post ops, value in vmm_d_weights and vmm_d_bias is re-loaded(init) each time.
-    Vmm vmm_index_in_y = vmm_d_weights;
-    Vmm vmm_index_in_x = vmm_d_bias;
-
-    Xbyak::Label l_table_constant;
-    Opmask k_mask = Xbyak::Opmask(1);
-
-    std::unordered_map<size_t, std::unique_ptr<jit_emitter>> emitters;
-
-    std::vector<size_t> store_pool_gpr_idxs;
-    std::vector<size_t> store_pool_vec_idxs;
-    std::vector<size_t> load_pool_gpr_idxs;
-
-    std::vector<std::shared_ptr<jit_uni_eltwise_injector_f32<isa>>> eltwise_injectors;
-    std::vector<std::shared_ptr<jit_uni_depthwise_injector_f32<isa>>> depthwise_injectors;
-    std::vector<std::shared_ptr<jit_uni_quantization_injector_f32<isa>>> quantization_injectors;
-
-    void emit_emitters_data() {
-        for (const auto& emitter : emitters) {
-            if (emitter.second)
-                emitter.second->emit_data();
-        }
-    }
-
-    inline void load(Xbyak::Reg64 reg_src, Vmm vmm_src, const int elt_num, const int offset = 0) {
-        emit_load(reg_src, vmm_src, jcp_.src_prc, Precision::FP32, elt_num, offset);
-    }
-
-    inline void load_weights(Xbyak::Reg64 reg_src, Vmm vmm_src, const int elt_num, const int offset = 0) {
-        emit_load(reg_src, vmm_src, Precision::FP32, Precision::FP32, elt_num, offset);
-    }
-
-    inline void emit_load(Xbyak::Reg64 reg_src, Vmm vmm_src, Precision src_prc, Precision dst_prc, const int elt_num, const int offset = 0) {
-        const auto seed = load_emitter_params(src_prc, dst_prc, elt_num).hash();
-        if (!emitters[seed]) {
-            emitters[seed].reset(new jit_load_emitter(this, isa, src_prc, dst_prc, elt_num));
-        }
-
-        emitters[seed]->emit_code({static_cast<size_t>(reg_src.getIdx()), static_cast<size_t>(offset)},
-                                  {static_cast<size_t>(vmm_src.getIdx())}, {}, {load_pool_gpr_idxs});
-    }
-
-    inline void store(Vmm vmm_dst, Xbyak::Reg64 reg_dst, const int elt_num, const int offset = 0) {
-        const auto seed = store_emitter_params(Precision::FP32, jcp_.dst_prc, elt_num).hash();
-        if (!emitters[seed]) {
-            emitters[seed].reset(new jit_store_emitter(this, isa, Precision::FP32, jcp_.dst_prc, elt_num));
-        }
-
-        // for cases when Store emitter need 2 aux vmm we can use vmm_dst as second aux vmm
-        std::vector<size_t> local_store_pool_vec_idxs = { static_cast<size_t>(vmm_dst.getIdx()) };
-        local_store_pool_vec_idxs.insert(local_store_pool_vec_idxs.begin(), store_pool_vec_idxs.begin(), store_pool_vec_idxs.end());
-
-        emitters[seed]->emit_code({static_cast<size_t>(vmm_dst.getIdx()), static_cast<size_t>(offset)},
-                                  {static_cast<size_t>(reg_dst.getIdx())},
-                                  {local_store_pool_vec_idxs}, {store_pool_gpr_idxs});
-    }
-
-    void nn_planar() {
-        Xbyak::Reg64 reg_index_h = reg_src_aux1;
-        Xbyak::Reg64 reg_index_w = reg_src_aux2;
-        mov(reg_index_h, reg_index);
-        // reg_index represent reg_index_w
-        add(reg_index, jcp_.OH * jcp_.indices_size);
-        // bk for reset to reg_index_w
-        mov(reg_index_w, reg_index);
-
-        Xbyak::Label out_loop_label;
-        Xbyak::Label out_loop_end;
-
-        Xbyak::Reg64 reg_work_amount_oh = rdi;
-        mov(reg_work_amount_oh, jcp_.OH);
-        L(out_loop_label);
-        {
-            // outloop status
-            cmp(reg_work_amount_oh, 1);
-            jl(out_loop_end, T_NEAR);
-
-            //reset work_amount to OW
-            mov(reg_work_amount, jcp_.OW);
-
-            Xbyak::Reg64 reg_src_h = rsi;
-            mov(reg_src_h, reg_src);
-            // index_h * IW * dataSize done when built to avoid redundent compute
-            mov(reg_index_offset, dword[reg_index_h]);
-            add(reg_src_h, reg_index_offset);  // reg_src_h now point to begin of row
-
-            // reset index_w, index_w * dataSize done when built to avoid redundent compute
-            mov(reg_index, reg_index_w);
-
-            Xbyak::Label nn_loop_label;
-            Xbyak::Label nn_loop_end_label;
-            Xbyak::Label nn_tail_loop_label;
-            Xbyak::Label nn_tail_loop_end_label;
-
-            L(nn_loop_label);   // inner loop
-            {
-                cmp(reg_work_amount, vector_step);
-                jl(nn_loop_end_label, T_NEAR);
-
-                uni_vmovdqu(vmm_index, ptr[reg_index]);
-                uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-                vgatherdps(vmm_val, ptr[reg_src_h + vmm_index], vmm_mask);
-                if (attr_.post_ops_.len() != 0)
-                    apply_post_ops(jcp_.dst_prc, 1);
-                store(vmm_val, reg_dst, vector_step);
-
-                add(reg_dst, vector_step * jcp_.dst_data_size);
-                add(reg_index, vector_step * jcp_.indices_size);
-                sub(reg_work_amount, vector_step);
-
-                jmp(nn_loop_label, T_NEAR);
-            }
-            L(nn_loop_end_label);
-
-            L(nn_tail_loop_label);
-            {
-                cmp(reg_work_amount, 1);
-                jl(nn_tail_loop_end_label, T_NEAR);
-
-                mov(reg_src_aux, reg_src_h);
-                mov(reg_index_offset, dword[reg_index]);
-                add(reg_src_aux, reg_index_offset);
-
-                load(reg_src_aux, vmm_val, scalar_step);
-                if (attr_.post_ops_.len() != 0)
-                    apply_post_ops(jcp_.dst_prc, 1);
-                store(vmm_val, reg_dst, scalar_step);
-
-                add(reg_dst, scalar_step * jcp_.dst_data_size);
-                add(reg_index, scalar_step * jcp_.indices_size);
-                sub(reg_work_amount, scalar_step);
-
-                jmp(nn_tail_loop_label, T_NEAR);
-            }
-            L(nn_tail_loop_end_label);    // inner loop end
-
-            //increment index_h to next row
-            add(reg_index_h, jcp_.indices_size);
-
-            sub(reg_work_amount_oh, 1);
-            jmp(out_loop_label, T_NEAR);
-        }
-        L(out_loop_end);
-    }
-
-    void nn_blk() {
-        Xbyak::Label nn_loop_label;
-        Xbyak::Label nn_loop_end_label;
-        L(nn_loop_label);
-        {
-            cmp(reg_work_amount, 0);
-            jle(nn_loop_end_label, T_NEAR);
-
-            mov(reg_src_aux, reg_src);
-            mov(reg_index_offset, dword[reg_index]);
-            add(reg_src_aux, reg_index_offset);
-
-            load(reg_src_aux, vmm_val, vector_step);
-            if (attr_.post_ops_.len() != 0)
-                apply_post_ops(jcp_.dst_prc, 0);
-            store(vmm_val, reg_dst, vector_step);
-            add(reg_dst, vector_step * jcp_.dst_data_size);
-
-            if (isa == cpu::x64::sse41) {
-                add(reg_src_aux, vector_step * jcp_.src_data_size);
-                load(reg_src_aux, vmm_val, vector_step);
-                if (attr_.post_ops_.len() != 0) {
-                    add(reg_oc_off, vector_step * sizeof(float));
-                    apply_post_ops(jcp_.dst_prc, 0);
-                    sub(reg_oc_off, vector_step * sizeof(float));
-                }
-                store(vmm_val, reg_dst, vector_step);
-                add(reg_dst, vector_step * jcp_.dst_data_size);
-            }
-
-            add(reg_index, jcp_.indices_size);
-            sub(reg_work_amount, 1);
-
-            jmp(nn_loop_label, T_NEAR);
-        }
-        L(nn_loop_end_label);
-    }
-
-    void nn_by_channel() {
-        // kernel for C * OW
-        Xbyak::Label out_loop_label;
-        Xbyak::Label out_loop_end;
-
-        Xbyak::Reg64 reg_work_amount_bk = reg_src_aux2;
-        Xbyak::Reg64 reg_oc_off_bk = rsi;
-        mov(reg_work_amount_bk, ptr[reg_params + GET_OFF(work_amount)]);
-        if (attr_.post_ops_.len() != 0) {
-            mov(reg_oc_off_bk, ptr[reg_params + GET_OFF(oc_off)]);
-        }
-
-        Xbyak::Reg64 reg_work_amount_out = reg_src_aux1;
-        mov(reg_work_amount_out, jcp_.OW);
-        L(out_loop_label);
-        {
-            cmp(reg_work_amount_out, 1);
-            jl(out_loop_end, T_NEAR);
-
-            //inner loop for C
-            Xbyak::Label nn_loop_label;
-            Xbyak::Label nn_loop_end_label;
-            Xbyak::Label nn_tail_loop_label;
-            Xbyak::Label nn_tail_loop_end_label;
-
-            // inner loop for C
-            // get current loop address reg_src_aux, from reg_src which is unchange, point this C * OW.
-            // reset offset and work_amount.
-            // dst and index address is continous, advanced each interator.
-            mov(reg_src_aux, reg_src);
-            // index*C*dataSize done when built to avoid redundent compute
-            mov(reg_index_offset, dword[reg_index]);
-            add(reg_src_aux, reg_index_offset);
-
-            mov(reg_work_amount, reg_work_amount_bk);
-            if (attr_.post_ops_.len() != 0)
-                mov(reg_oc_off, reg_oc_off_bk);
-
-            L(nn_loop_label);
-            {
-                cmp(reg_work_amount, vector_step);
-                jl(nn_loop_end_label, T_NEAR);
-
-                load(reg_src_aux, vmm_val, vector_step);
-                if (attr_.post_ops_.len() != 0)
-                    apply_post_ops(jcp_.dst_prc, 0);
-                store(vmm_val, reg_dst, vector_step);
-
-                add(reg_dst, vector_step * jcp_.dst_data_size);
-                add(reg_src_aux, vector_step * jcp_.src_data_size);
-                add(reg_oc_off, vector_step * sizeof(float));
-                sub(reg_work_amount, vector_step);
-
-                jmp(nn_loop_label, T_NEAR);
-            }
-            L(nn_loop_end_label);
-
-            if (tail_step != 0) {
-                load(reg_src_aux, vmm_val, tail_step);
-                if (attr_.post_ops_.len() != 0)
-                    apply_post_ops(jcp_.dst_prc, 0);
-                store(vmm_val, reg_dst, tail_step);
-
-                // check to remove below
-                add(reg_dst, tail_step * jcp_.dst_data_size);
-                add(reg_src_aux, tail_step * jcp_.src_data_size);
-                add(reg_oc_off, tail_step * sizeof(float));
-                sub(reg_work_amount, tail_step);
-            }
-            add(reg_index, jcp_.indices_size);
-            sub(reg_work_amount_out, 1);
-            jmp(out_loop_label, T_NEAR);
-        }
-        L(out_loop_end);
-    }
-
-    void linear_onnx_c_gathered() {
-        mov(reg_dst, ptr[reg_params + GET_OFF(dst)]);
-        // load weight
-        mov(reg_src, ptr[reg_params + GET_OFF(weight_ptr[0])]);
-        mov(reg_src_aux, ptr[reg_params + GET_OFF(weight_ptr[0]) + sizeof(size_t)]);
-        uni_vbroadcastss(vmm_weightL, ptr[reg_src]);
-        uni_vbroadcastss(vmm_weightR, ptr[reg_src_aux]);
-        if (jcp_.spatial_dim_size > 1) {
-            mov(reg_src_aux1, ptr[reg_params + GET_OFF(weight_ptr[0]) + 2 * sizeof(size_t)]);
-            mov(reg_src_aux2, ptr[reg_params + GET_OFF(weight_ptr[0]) + 3 * sizeof(size_t)]);
-            uni_vbroadcastss(vmm_weightT, ptr[reg_src_aux1]);
-            uni_vbroadcastss(vmm_weightB, ptr[reg_src_aux2]);
-        }
-        if (jcp_.spatial_dim_size > 2) {
-            mov(reg_src, ptr[reg_params + GET_OFF(weight_ptr[0]) + 4 * sizeof(size_t)]);
-            mov(reg_src_aux, ptr[reg_params + GET_OFF(weight_ptr[0]) + 5 * sizeof(size_t)]);
-            uni_vbroadcastss(vmm_weightF, ptr[reg_src]);
-            uni_vbroadcastss(vmm_weightE, ptr[reg_src_aux]);
-        }
-        // load src
-        mov(reg_src, ptr[reg_params + GET_OFF(src_ptr[0])]);
-        mov(reg_src_aux, ptr[reg_params + GET_OFF(src_ptr[0]) + sizeof(size_t)]);
-        if (jcp_.spatial_dim_size > 1) {
-            mov(reg_src_aux1, ptr[reg_params + GET_OFF(src_ptr[0]) + 2 * sizeof(size_t)]);
-            mov(reg_src_aux2, ptr[reg_params + GET_OFF(src_ptr[0]) + 3 * sizeof(size_t)]);
-        }
-        Xbyak::Reg64 reg_src_aux4 = r14;
-        Xbyak::Reg64 reg_src_aux5 = rdx;
-        Xbyak::Reg64 reg_src_aux6 = rsi;
-        Xbyak::Reg64 reg_src_aux7 = rbp;
-        if (jcp_.spatial_dim_size > 2) {
-            mov(reg_src_aux4, ptr[reg_params + GET_OFF(src_ptr[0]) + 4 * sizeof(size_t)]);
-            mov(reg_src_aux5, ptr[reg_params + GET_OFF(src_ptr[0]) + 5 * sizeof(size_t)]);
-            mov(reg_src_aux6, ptr[reg_params + GET_OFF(src_ptr[0]) + 6 * sizeof(size_t)]);
-            mov(reg_src_aux7, ptr[reg_params + GET_OFF(src_ptr[0]) + 7 * sizeof(size_t)]);
-        }
-        mov(reg_work_amount, ptr[reg_params + GET_OFF(work_amount)]);
-
-        int blk = (isa == cpu::x64::sse41) ? (2 * vector_step) : vector_step;
-        int dst_stride = (jcp_.layout == InterpolateLayoutType::by_channel) ? (vector_step * jcp_.dst_data_size) :
-                                            (blk * jcp_.OW * jcp_.OH * jcp_.OD * jcp_.dst_data_size);
-        int src_stride = (jcp_.layout == InterpolateLayoutType::by_channel) ? (vector_step * jcp_.src_data_size) :
-                                            (blk * jcp_.IW * jcp_.IH * jcp_.ID * jcp_.src_data_size);
-
-        Xbyak::Label main_loop_label;
-        Xbyak::Label main_loop_end_label;
-        Xbyak::Label blk_tail_loop_label;
-        Xbyak::Label blk_tail_loop_end_label;
-        Xbyak::Label tail_loop_label;
-        Xbyak::Label tail_loop_end_label;
-        L(main_loop_label);
-        {
-            if (jcp_.layout == InterpolateLayoutType::by_channel) {
-                cmp(reg_work_amount, vector_step);
-                jl(main_loop_end_label, T_NEAR);
-            } else {
-                cmp(reg_work_amount, 1);
-                jl(main_loop_end_label, T_NEAR);
-            }
-            // progressive manner
-            load(reg_src, vmm_valTL, vector_step);
-            load(reg_src_aux, vmm_valTR, vector_step);
-            if (jcp_.spatial_dim_size == 1) {
-                linear_onnx_worker_1d();
-            }
-            if (jcp_.spatial_dim_size > 1) {
-                load(reg_src_aux1, vmm_valBL, vector_step);
-                load(reg_src_aux2, vmm_valBR, vector_step);
-                linear_onnx_worker_2d();
-            }
-            if (jcp_.spatial_dim_size > 2) {
-                uni_vmovups(vmm_d_bias, vmm_valTR);  // temporally save front result to temp_vmm
-                load(reg_src_aux4, vmm_valTL, vector_step);
-                load(reg_src_aux5, vmm_valTR, vector_step);
-                load(reg_src_aux6, vmm_valBL, vector_step);
-                load(reg_src_aux7, vmm_valBR, vector_step);
-
-                // 2d for end depth
-                linear_onnx_worker_2d();
-                // 3th dimension
-                uni_vmulps(vmm_valTR, vmm_valTR, vmm_weightE); // end_value * end_weight
-                uni_vfmadd231ps(vmm_valTR, vmm_d_bias, vmm_weightF); // start_value * start_weight + end_value * end_weight
-            }
-
-            if (attr_.post_ops_.len() != 0) {
-                apply_post_ops(jcp_.dst_prc, false);  // vmm_val is vmm_valTR
-                add(reg_oc_off, vector_step * sizeof(float));
-            }
-            store(vmm_valTR, reg_dst, vector_step);
-
-            if ((isa == cpu::x64::sse41) && (jcp_.layout == InterpolateLayoutType::block)) {
-                int offset_src = vector_step * jcp_.src_data_size;
-                load(reg_src, vmm_valTL, vector_step, offset_src);
-                load(reg_src_aux, vmm_valTR, vector_step, offset_src);
-                if (jcp_.spatial_dim_size == 1) {
-                    linear_onnx_worker_1d();
-                }
-                if (jcp_.spatial_dim_size > 1) {
-                    load(reg_src_aux1, vmm_valBL, vector_step, offset_src);
-                    load(reg_src_aux2, vmm_valBR, vector_step, offset_src);
-                    linear_onnx_worker_2d();
-                }
-                if (jcp_.spatial_dim_size > 2) {
-                    uni_vmovups(vmm_d_bias, vmm_valTR);  // temporally save front result to temp_vmm
-                    load(reg_src_aux4, vmm_valTL, vector_step, offset_src);
-                    load(reg_src_aux5, vmm_valTR, vector_step, offset_src);
-                    load(reg_src_aux6, vmm_valBL, vector_step, offset_src);
-                    load(reg_src_aux7, vmm_valBR, vector_step, offset_src);
-                    // 2d for end depth
-                    linear_onnx_worker_2d();
-                    // 3th dimension
-                    uni_vmulps(vmm_valTR, vmm_valTR, vmm_weightE); // end_value * end_weight
-                    uni_vfmadd231ps(vmm_valTR, vmm_d_bias, vmm_weightF); // start_value * start_weight + end_value * end_weight
-                }
-
-                if (attr_.post_ops_.len() != 0) {
-                    apply_post_ops(jcp_.dst_prc, false);
-                    add(reg_oc_off, vector_step * sizeof(float));
-                }
-                int offset_dst = vector_step * jcp_.dst_data_size;
-                store(vmm_valTR, reg_dst, vector_step, offset_dst);
-            }
-            add(reg_dst, dst_stride);
-            add(reg_src, src_stride);
-            add(reg_src_aux, src_stride);
-            if (jcp_.spatial_dim_size > 1) {
-                add(reg_src_aux1, src_stride);
-                add(reg_src_aux2, src_stride);
-            }
-            if (jcp_.spatial_dim_size > 2) {
-                add(reg_src_aux4, src_stride);
-                add(reg_src_aux5, src_stride);
-                add(reg_src_aux6, src_stride);
-                add(reg_src_aux7, src_stride);
-            }
-            if (jcp_.layout == InterpolateLayoutType::by_channel) {
-                sub(reg_work_amount, vector_step);    // work_amount is c
-            } else {
-                sub(reg_work_amount, 1);       // work_amount = div_up(c, blk), no tails
-            }
-
-            jmp(main_loop_label, T_NEAR);
-        }
-        L(main_loop_end_label);
-
-        if ((jcp_.layout == InterpolateLayoutType::by_channel) && (tail_step != 0)) {
-            load(reg_src, vmm_valTL, tail_step);
-            load(reg_src_aux, vmm_valTR, tail_step);
-            if (jcp_.spatial_dim_size == 1) {
-                linear_onnx_worker_1d();
-            }
-            if (jcp_.spatial_dim_size > 1) {
-                load(reg_src_aux1, vmm_valBL, tail_step);
-                load(reg_src_aux2, vmm_valBR, tail_step);
-                linear_onnx_worker_2d();
-            }
-            if (jcp_.spatial_dim_size > 2) {
-                uni_vmovups(vmm_d_bias, vmm_valTR);  // temporally save front result to temp_vmm
-
-                load(reg_src_aux4, vmm_valTL, tail_step);
-                load(reg_src_aux5, vmm_valTR, tail_step);
-                load(reg_src_aux6, vmm_valBL, tail_step);
-                load(reg_src_aux7, vmm_valBR, tail_step);
-                // 2d for end depth
-                linear_onnx_worker_2d();
-                // 3th dimension
-                uni_vmulps(vmm_valTR, vmm_valTR, vmm_weightE); // end_value * end_weight
-                uni_vfmadd231ps(vmm_valTR, vmm_d_bias, vmm_weightF); // start_value * start_weight + end_value * end_weight
-            }
-
-            if (attr_.post_ops_.len() != 0) {
-                apply_post_ops(jcp_.dst_prc, false);  // vmm_val is vmm_valTR
-                add(reg_oc_off, tail_step * sizeof(float));
-            }
-
-            store(vmm_valTR, reg_dst, tail_step);
-        }
-    }
-
-    void linear_onnx_planar() {
-        mov(reg_dst, ptr[reg_params + GET_OFF(dst)]);
-        mov(reg_src, ptr[reg_params + GET_OFF(src_ptr[0])]);
-        mov(reg_index, ptr[reg_params + GET_OFF(index)]);
-        mov(reg_src_aux, ptr[reg_params + GET_OFF(weight_ptr[0])]);
-        mov(reg_work_amount, ptr[reg_params + GET_OFF(work_amount)]);
-
-        int index_stride = jcp_.OW * jcp_.OH * jcp_.OD * jcp_.indices_size;
-        int weight_stride = jcp_.OW * jcp_.OH * jcp_.OD * sizeof(float);
-
-        Xbyak::Label main_loop_label;
-        Xbyak::Label main_loop_end_label;
-        Xbyak::Label tail_loop_label;
-        Xbyak::Label tail_loop_end_label;
-        L(main_loop_label);
-        {
-            cmp(reg_work_amount, vector_step);
-            jl(main_loop_end_label, T_NEAR);
-
-            uni_vmovdqu(vmm_index, ptr[reg_index]);
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            vgatherdps(vmm_valTL, ptr[reg_src + vmm_index], vmm_mask);
-
-            uni_vmovdqu(vmm_index, ptr[reg_index + index_stride]);
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            vgatherdps(vmm_valTR, ptr[reg_src + vmm_index], vmm_mask);
-
-            load_weights(reg_src_aux, vmm_weightL, vector_step);
-            load_weights(reg_src_aux, vmm_weightR, vector_step, weight_stride);
-
-            // progressive manner
-            if (jcp_.spatial_dim_size == 1) {
-                linear_onnx_worker_1d();
-            }
-            if (jcp_.spatial_dim_size > 1) {
-                uni_vmovdqu(vmm_index, ptr[reg_index + 2 * index_stride]);
-                uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-                vgatherdps(vmm_valBL, ptr[reg_src + vmm_index], vmm_mask);
-
-                uni_vmovdqu(vmm_index, ptr[reg_index + 3 * index_stride]);
-                uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-                vgatherdps(vmm_valBR, ptr[reg_src + vmm_index], vmm_mask);
-
-                load_weights(reg_src_aux, vmm_weightT, vector_step, 2 * weight_stride);
-                load_weights(reg_src_aux, vmm_weightB, vector_step, 3 * weight_stride);
-
-                linear_onnx_worker_2d();
-            }
-            if (jcp_.spatial_dim_size > 2) {
-                uni_vmovups(vmm_d_bias, vmm_valTR);  // temporally save front result to temp_vmm
-
-                // for end depth
-                uni_vmovdqu(vmm_index, ptr[reg_index + 4 * index_stride]);
-                uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-                vgatherdps(vmm_valTL, ptr[reg_src + vmm_index], vmm_mask);
-
-                uni_vmovdqu(vmm_index, ptr[reg_index + 5 * index_stride]);
-                uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-                vgatherdps(vmm_valTR, ptr[reg_src + vmm_index], vmm_mask);
-
-                uni_vmovdqu(vmm_index, ptr[reg_index + 6 * index_stride]);
-                uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-                vgatherdps(vmm_valBL, ptr[reg_src + vmm_index], vmm_mask);
-
-                uni_vmovdqu(vmm_index, ptr[reg_index + 7 * index_stride]);
-                uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-                vgatherdps(vmm_valBR, ptr[reg_src + vmm_index], vmm_mask);
-
-                linear_onnx_worker_2d();
-
-                load_weights(reg_src_aux, vmm_weightE, vector_step, 5 * weight_stride);
-                load_weights(reg_src_aux, vmm_weightF, vector_step, 4 * weight_stride);
-
-                uni_vmulps(vmm_valTR, vmm_valTR, vmm_weightE); // end_value * end_weight
-                uni_vfmadd231ps(vmm_valTR, vmm_d_bias, vmm_weightF); // start_value * start_weight + end_value * end_weight
-            }
-
-            if (attr_.post_ops_.len() != 0) {
-                apply_post_ops(jcp_.dst_prc, true);  // vmm_val is vmm_valTR, broadcase is true
-            }
-            store(vmm_valTR, reg_dst, vector_step);
-
-            add(reg_dst, vector_step * jcp_.dst_data_size);
-            add(reg_src_aux, vector_step * sizeof(float));
-            add(reg_index, vector_step * jcp_.indices_size);
-            sub(reg_work_amount, vector_step);
-
-            jmp(main_loop_label, T_NEAR);
-        }
-        L(main_loop_end_label);
-
-        L(tail_loop_label);
-        {
-            cmp(reg_work_amount, 1);
-            jl(tail_loop_end_label, T_NEAR);
-
-            mov(reg_src_aux1, reg_src);
-            mov(reg_index_offset, dword[reg_index]);
-            add(reg_src_aux1, reg_index_offset);
-            load(reg_src_aux1, vmm_valTL, scalar_step);
-
-            mov(reg_src_aux1, reg_src);
-            mov(reg_index_offset, dword[reg_index + index_stride]);
-            add(reg_src_aux1, reg_index_offset);
-            load(reg_src_aux1, vmm_valTR, scalar_step);
-
-            load_weights(reg_src_aux, vmm_weightL, scalar_step, 0);
-            load_weights(reg_src_aux, vmm_weightR, scalar_step, weight_stride);
-
-            if (jcp_.spatial_dim_size == 1) {
-                linear_onnx_worker_1d();
-            }
-            if (jcp_.spatial_dim_size > 1) {
-                mov(reg_src_aux1, reg_src);
-                mov(reg_index_offset, dword[reg_index + 2 * index_stride]);
-                add(reg_src_aux1, reg_index_offset);
-                load(reg_src_aux1, vmm_valBL, scalar_step);
-
-                mov(reg_src_aux1, reg_src);
-                mov(reg_index_offset, dword[reg_index + 3 * index_stride]);
-                add(reg_src_aux1, reg_index_offset);
-                load(reg_src_aux1, vmm_valBR, scalar_step);
-
-                load_weights(reg_src_aux, vmm_weightT, scalar_step, 2 * weight_stride);
-                load_weights(reg_src_aux, vmm_weightB, scalar_step, 3 * weight_stride);
-
-                linear_onnx_worker_2d();
-            }
-            if (jcp_.spatial_dim_size > 2) {
-                uni_vmovups(vmm_d_bias, vmm_valTR);  // save from front result to temp_vmm
-
-                // for end depth
-                mov(reg_src_aux1, reg_src);
-                mov(reg_index_offset, dword[reg_index + 4 * index_stride]);
-                add(reg_src_aux1, reg_index_offset);
-                load(reg_src_aux1, vmm_valTL, scalar_step);
-
-                mov(reg_src_aux1, reg_src);
-                mov(reg_index_offset, dword[reg_index + 5 * index_stride]);
-                add(reg_src_aux1, reg_index_offset);
-                load(reg_src_aux1, vmm_valTR, scalar_step);
-
-                mov(reg_src_aux1, reg_src);
-                mov(reg_index_offset, dword[reg_index + 6 * index_stride]);
-                add(reg_src_aux1, reg_index_offset);
-                load(reg_src_aux1, vmm_valBL, scalar_step);
-
-                mov(reg_src_aux1, reg_src);
-                mov(reg_index_offset, dword[reg_index + 7 * index_stride]);
-                add(reg_src_aux1, reg_index_offset);
-                load(reg_src_aux1, vmm_valBR, scalar_step);
-
-                linear_onnx_worker_2d();
-
-                load_weights(reg_src_aux, vmm_weightE, scalar_step, 5 * weight_stride);
-                load_weights(reg_src_aux, vmm_weightF, scalar_step, 4 * weight_stride);
-
-                uni_vmulps(vmm_valTR, vmm_valTR, vmm_weightE); // end_value * end_weight
-                uni_vfmadd231ps(vmm_valTR, vmm_d_bias, vmm_weightF); // start_value * start_weight + end_value * end_weight
-            }
-
-            if (attr_.post_ops_.len() != 0) {
-                apply_post_ops(jcp_.dst_prc, true);  // process on vmm_val, vmm_val is vmm_valTR, and bc
-            }
-            store(vmm_valTR, reg_dst, scalar_step);
-
-            add(reg_dst, scalar_step * jcp_.dst_data_size);
-            add(reg_src_aux, scalar_step * sizeof(float));
-            add(reg_index, scalar_step * jcp_.indices_size);
-            sub(reg_work_amount, scalar_step);
-
-            jmp(tail_loop_label, T_NEAR);
-        }
-        L(tail_loop_end_label);
-    }
-
-    inline void linear_onnx_worker_1d() {
-        uni_vmulps(vmm_valTR, vmm_valTR, vmm_weightR);
-        uni_vfmadd231ps(vmm_valTR, vmm_valTL, vmm_weightL);
-    }
-
-    // weightT * (srcTL * weightL + srcTR * weightR) +
-    // weightB * (srcBL * weightL + srcBR * weightR)
-    inline void linear_onnx_worker_2d() {
-        uni_vmulps(vmm_valTR, vmm_valTR, vmm_weightR);
-        uni_vmulps(vmm_valBR, vmm_valBR, vmm_weightR);
-        uni_vfmadd231ps(vmm_valTR, vmm_valTL, vmm_weightL);
-        uni_vfmadd231ps(vmm_valBR, vmm_valBL, vmm_weightL);
-        uni_vmulps(vmm_valTR, vmm_valTR, vmm_weightT);
-        uni_vfmadd231ps(vmm_valTR, vmm_valBR, vmm_weightB);
-    }
-
-    void cubic_c_gathered() {
-        mov(reg_dst, ptr[reg_params + GET_OFF(dst)]);
-        mov(reg_src, ptr[reg_params + GET_OFF(src_ptr[0])]);
-        mov(reg_index, ptr[reg_params + GET_OFF(index)]);
-        mov(reg_work_amount, ptr[reg_params + GET_OFF(work_amount)]);
-
-        // weight_ptr[0] point to weightX
-        mov(reg_src_aux1, ptr[reg_params + GET_OFF(weight_ptr[0])]);
-        uni_vbroadcastss(vmm_weightX0, ptr[reg_src_aux1]);
-        uni_vbroadcastss(vmm_weightX1, ptr[reg_src_aux1 + 1 * sizeof(float)]);
-        uni_vbroadcastss(vmm_weightX2, ptr[reg_src_aux1 + 2 * sizeof(float)]);
-        uni_vbroadcastss(vmm_weightX3, ptr[reg_src_aux1 + 3 * sizeof(float)]);
-
-        // weight_ptr[1] point to weightY
-        mov(reg_src_aux1, ptr[reg_params + GET_OFF(weight_ptr[0]) + sizeof(size_t)]);
-        uni_vbroadcastss(vmm_weightY0, ptr[reg_src_aux1]);
-        uni_vbroadcastss(vmm_weightY1, ptr[reg_src_aux1 + 1 * sizeof(float)]);
-        uni_vbroadcastss(vmm_weightY2, ptr[reg_src_aux1 + 2 * sizeof(float)]);
-        uni_vbroadcastss(vmm_weightY3, ptr[reg_src_aux1 + 3 * sizeof(float)]);
-
-        int blk = (isa == cpu::x64::sse41) ? (2 * vector_step) : vector_step;
-
-        Xbyak::Label main_loop_label;
-        Xbyak::Label main_loop_end_label;
-        Xbyak::Label tail_loop_label;
-        Xbyak::Label tail_loop_end_label;
-        L(main_loop_label);
-        {
-            if (jcp_.layout == InterpolateLayoutType::by_channel) {
-                cmp(reg_work_amount, vector_step);
-                jl(main_loop_end_label, T_NEAR);
-            } else {
-                cmp(reg_work_amount, 1);
-                jl(tail_loop_end_label, T_NEAR);
-            }
-
-            uni_vpxor(vmm_val, vmm_val, vmm_val);
-
-            cubic_c_gathered_matrix(false);
-
-            if (attr_.post_ops_.len() != 0) {
-                apply_post_ops(jcp_.dst_prc, false);     // vmm_val is default dst value to post_ops and store
-                add(reg_oc_off, vector_step * sizeof(float));
-            }
-            store(vmm_val, reg_dst, vector_step);
-
-            if ((isa == cpu::x64::sse41) && (jcp_.layout == InterpolateLayoutType::block)) {
-                // vmm is xmm here
-                add(reg_src, vector_step * jcp_.src_data_size);
-                add(reg_dst, vector_step * jcp_.dst_data_size);
-
-                uni_vpxor(vmm_val, vmm_val, vmm_val);
-
-                cubic_c_gathered_matrix(false);
-
-                if (attr_.post_ops_.len() != 0) {
-                    apply_post_ops(jcp_.dst_prc, false);
-                    add(reg_oc_off, vector_step * sizeof(float));  // second vector_step for one blk
-                }
-                store(vmm_val, reg_dst, vector_step);
-
-                sub(reg_src, vector_step * jcp_.src_data_size);
-                sub(reg_dst, vector_step * jcp_.dst_data_size);
-            }
-            if (jcp_.layout == InterpolateLayoutType::by_channel) {
-                int dst_stride = vector_step * jcp_.dst_data_size;
-                int src_stride = vector_step * jcp_.src_data_size;
-                add(reg_dst, dst_stride);
-                add(reg_src, src_stride);
-                sub(reg_work_amount, vector_step);    // work_amount is c
-            } else {
-                int dst_stride = blk * jcp_.OW * jcp_.OH * jcp_.dst_data_size;
-                int src_stride = blk * jcp_.IW * jcp_.IH * jcp_.src_data_size;
-                add(reg_dst, dst_stride);
-                add(reg_src, src_stride);
-                sub(reg_work_amount, 1);  // work_amount = div_up(c, blk), no tails
-            }
-
-            jmp(main_loop_label, T_NEAR);
-        }
-        L(main_loop_end_label);
-
-        // only for by_channel layout for tails.
-        L(tail_loop_label);
-        {
-            cmp(reg_work_amount, 1);
-            jl(tail_loop_end_label, T_NEAR);
-
-            // store final computed value
-            uni_vpxor(vmm_val, vmm_val, vmm_val);
-
-            cubic_c_gathered_matrix(true);
-
-            if (attr_.post_ops_.len() != 0) {
-                apply_post_ops(jcp_.dst_prc, false);     // vmm_val is default dst value
-                add(reg_oc_off, scalar_step * sizeof(float));
-            }
-            store(vmm_val, reg_dst, scalar_step);
-
-            int dst_stride = scalar_step * jcp_.dst_data_size;
-            int src_stride = scalar_step * jcp_.src_data_size;
-            add(reg_dst, dst_stride);
-            add(reg_src, src_stride);
-            sub(reg_work_amount, scalar_step);    // work_amount is c
-
-            jmp(tail_loop_label, T_NEAR);
-        }
-        L(tail_loop_end_label);
-    }
-
-    inline void cubic_c_gathered_matrix(bool is_scalar) {
-        // y0:  (x0 * weightX0 + x1 * weightX1 + x2 * weightX2 + x3 * weightX3) * weightY0
-        cubic_c_gathered_line(0, vmm_weightY0, is_scalar);
-        // y1
-        cubic_c_gathered_line(4, vmm_weightY1, is_scalar);
-        // y2
-        cubic_c_gathered_line(8, vmm_weightY2, is_scalar);
-        // y3
-        cubic_c_gathered_line(12, vmm_weightY3, is_scalar);
-    }
-
-    inline void cubic_c_gathered_line(int index_start, Vmm vmm_weight, bool is_scalar) {
-        uni_vpxor(vmm_dstX, vmm_dstX, vmm_dstX);
-        cubic_c_gathered_pixel(index_start, vmm_weightX0, is_scalar);
-        cubic_c_gathered_pixel(index_start + 1, vmm_weightX1, is_scalar);
-        cubic_c_gathered_pixel(index_start + 2, vmm_weightX2, is_scalar);
-        cubic_c_gathered_pixel(index_start + 3, vmm_weightX3, is_scalar);
-        uni_vfmadd231ps(vmm_val, vmm_dstX, vmm_weight);
-    }
-
-    inline void cubic_c_gathered_pixel(int i, Vmm vmm_weight, bool is_scalar) {
-        mov(reg_src_aux, reg_src);
-        mov(reg_index_offset, dword[reg_index + i * jcp_.indices_size]);
-        add(reg_src_aux, reg_index_offset);
-        int step = is_scalar ? 1 : vlen / sizeof(float);
-        load(reg_src_aux, vmm_src, step);
-        uni_vfmadd231ps(vmm_dstX, vmm_src, vmm_weight);
-    }
-
-    void cubic_planar() {
-        mov(reg_table, l_table_constant);
-        // src_ptr[2] for oh sequence, src_ptr[3] for ow sequence
-        mov(reg_tbl_y, ptr[reg_params + GET_OFF(src_ptr[0]) + 2 * sizeof(size_t)]);
-        mov(reg_tbl_x, ptr[reg_params + GET_OFF(src_ptr[0]) + 3 * sizeof(size_t)]);
-        uni_vmovdqu(vmm_one, cubic_planar_table_val(0));
-        uni_vpxor(vmm_zero, vmm_zero, vmm_zero);
-
-        mov(reg_dst, ptr[reg_params + GET_OFF(dst)]);
-        mov(reg_src, ptr[reg_params + GET_OFF(src_ptr[0])]);
-        // index_OW
-        mov(reg_index, ptr[reg_params + GET_OFF(index)]);
-        // index_OH from src_ptr[1]
-        Xbyak::Reg64 reg_index_y = reg_src_aux;
-        mov(reg_index_y, ptr[reg_params + GET_OFF(src_ptr[0]) + sizeof(size_t)]);
-        // weight_OW
-        Xbyak::Reg64 reg_weight_x = reg_src_aux1;
-        mov(reg_weight_x, ptr[reg_params + GET_OFF(weight_ptr[0])]);
-        // weight_OH
-        Xbyak::Reg64 reg_weight_y = reg_src_aux2;
-        mov(reg_weight_y, ptr[reg_params + GET_OFF(weight_ptr[0]) + sizeof(size_t)]);
-        mov(reg_work_amount, ptr[reg_params + GET_OFF(work_amount)]);
-
-        int grid_len = 4;
-
-        // 0   1   2   3   4   5   6   7   8   9   10   11   12   13   14   15   16   17   18   19
-        // 20  21  22  23  24  25  26  27  28  29  30   31   32   33   34   35   36   37   38   39
-        // for 3th step(8): 16  17  18  19  20  21  22  23
-        //               y: 0   0   0   0   1   1   1   1
-        //               x: 16  17  18  19  0   1   2   3
-
-        Xbyak::Label main_loop_label;
-        Xbyak::Label main_loop_end_label;
-        Xbyak::Label tail_loop_label;
-        Xbyak::Label tail_loop_end_label;
-        L(main_loop_label);
-        {
-            cmp(reg_work_amount, vector_step);
-            jl(main_loop_end_label, T_NEAR);
-
-            // vmm_tbl_y: (0 0 0 0 1 1 1 1 * index_size) --> (0 0 0 0 4 4 4 4)
-            uni_vmovdqu(vmm_tbl_y, ptr[reg_tbl_y]);
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            // vmm_index_in_y: 0 0 0 0 2 2 2 2
-            vpgatherdd(vmm_index_in_y, ptr[reg_index_y + vmm_tbl_y], vmm_mask);
-
-            // use vmm_val temporally for value in reg_tbl_x: 16  17  18  19  0   1   2   3
-            uni_vmovdqu(vmm_val, ptr[reg_tbl_x]);
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            // e.g. vmm_index_in_x: 32 34 36 38 0 2 4 6, now save src index.
-            vpgatherdd(vmm_index_in_x, ptr[reg_index + vmm_val], vmm_mask);
-
-            // build weightX used in y0-y3
-            // weight format: w0_0 w1_0 w2_0 w3_0 w0_1 w1_1 w2_1 w3_1 ...
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            vgatherdps(vmm_weightX0, ptr[reg_weight_x + vmm_val * grid_len], vmm_mask);  // 4 in vmm_val for weight_size, another 4 for grid_len
-
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            // shift weight_size then gather second weight
-            vgatherdps(vmm_weightX1, ptr[reg_weight_x + sizeof(float) + (vmm_val * grid_len)], vmm_mask);
-
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            vgatherdps(vmm_weightX2, ptr[reg_weight_x + 2 * sizeof(float) + (vmm_val * grid_len)], vmm_mask);
-
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            vgatherdps(vmm_weightX3, ptr[reg_weight_x + 3 * sizeof(float) + (vmm_val * grid_len)], vmm_mask);
-            // vmm_val is now relieved and used for dst_value
-
-            uni_vpxor(vmm_val, vmm_val, vmm_val);
-            // y0
-            vpsubd(vmm_index_y_itr, vmm_index_in_y, vmm_one);
-            // crop to [0, IH - 1]
-            vpminsd(vmm_index_y_itr, vmm_index_y_itr, cubic_planar_table_val(1));
-            vpmaxsd(vmm_index_y_itr, vmm_index_y_itr, vmm_zero);
-
-            // weight y0
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            vgatherdps(vmm_weightY, ptr[reg_weight_y + (vmm_tbl_y * grid_len)], vmm_mask);
-            cubic_planar_line(false);
-
-            // y1
-            // crop to [0, IH - 1]
-            vpminsd(vmm_index_y_itr, vmm_index_in_y, cubic_planar_table_val(1));
-            vpmaxsd(vmm_index_y_itr, vmm_index_y_itr, vmm_zero);
-            // weight y1: shift weight_size
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            vgatherdps(vmm_weightY, ptr[reg_weight_y + sizeof(float) + (vmm_tbl_y * grid_len)], vmm_mask);
-            cubic_planar_line(false);
-
-            // y2
-            vpaddd(vmm_index_y_itr, vmm_index_in_y, vmm_one);
-            // crop to [0, IH - 1]
-            vpminsd(vmm_index_y_itr, vmm_index_y_itr, cubic_planar_table_val(1));
-            vpmaxsd(vmm_index_y_itr, vmm_index_y_itr, vmm_zero);
-            // weight y2
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            vgatherdps(vmm_weightY, ptr[reg_weight_y + 2 * sizeof(float) + (vmm_tbl_y * grid_len)], vmm_mask);
-            cubic_planar_line(false);
-
-            // y3
-            vpaddd(vmm_index_y_itr, vmm_index_in_y, vmm_one);
-            vpaddd(vmm_index_y_itr, vmm_index_y_itr, vmm_one);
-            // crop to [0, IH - 1]
-            vpminsd(vmm_index_y_itr, vmm_index_y_itr, cubic_planar_table_val(1));
-            vpmaxsd(vmm_index_y_itr, vmm_index_y_itr, vmm_zero);
-            // weight y3
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            vgatherdps(vmm_weightY, ptr[reg_weight_y + 3 * sizeof(float) + (vmm_tbl_y * grid_len)], vmm_mask);
-            cubic_planar_line(false);
-
-            if (attr_.post_ops_.len() != 0) {
-                apply_post_ops(jcp_.dst_prc, true);  // oc_off is broadcast and always the same value for this channel
-            }
-            store(vmm_val, reg_dst, vector_step);
-
-            add(reg_tbl_y, vector_step * sizeof(int));  // sizeof(int): sequence by dd()
-            add(reg_tbl_x, vector_step * sizeof(int));
-            add(reg_dst, vector_step * jcp_.dst_data_size);
-
-            sub(reg_work_amount, vector_step);
-
-            jmp(main_loop_label, T_NEAR);
-        }
-        L(main_loop_end_label);
-
-        L(tail_loop_label);
-        {
-            cmp(reg_work_amount, 1);
-            jl(tail_loop_end_label, T_NEAR);
-
-            // get idx for input
-            uni_vmovss(Xmm(vmm_tbl_y.getIdx()), ptr[reg_tbl_y]);
-            gather_i32_indices(vmm_index_in_y, reg_index_y, 0, vmm_tbl_y, 1, Precision::I32, true);
-
-            uni_vmovss(Xmm(vmm_val.getIdx()), ptr[reg_tbl_x]);
-            gather_i32_indices(vmm_index_in_x, reg_index, 0, vmm_val, 1, Precision::I32, true);
-            // gather weightX by input idx, used in y0-y3
-            gather_i32_indices(vmm_weightX0, reg_weight_x, 0, vmm_val, grid_len, Precision::FP32, true);
-            gather_i32_indices(vmm_weightX1, reg_weight_x, sizeof(float), vmm_val, grid_len, Precision::FP32, true);
-            gather_i32_indices(vmm_weightX2, reg_weight_x, 2 * sizeof(float), vmm_val, grid_len, Precision::FP32, true);
-            gather_i32_indices(vmm_weightX3, reg_weight_x, 3 * sizeof(float), vmm_val, grid_len, Precision::FP32, true);
-            // vmm_val is now relieved and used for dst_value
-
-            uni_vpxor(vmm_val, vmm_val, vmm_val);
-            // y0
-            vpsubd(vmm_index_y_itr, vmm_index_in_y, vmm_one);
-            // crop to [0, IH - 1]
-            vpminsd(vmm_index_y_itr, vmm_index_y_itr, cubic_planar_table_val(1));
-            vpmaxsd(vmm_index_y_itr, vmm_index_y_itr, vmm_zero);
-
-            gather_i32_indices(vmm_weightY, reg_weight_y, 0, vmm_tbl_y, grid_len, Precision::FP32, true);
-            cubic_planar_line(true);
-
-            // y1
-            // crop to [0, IH - 1]
-            vpminsd(vmm_index_y_itr, vmm_index_in_y, cubic_planar_table_val(1));
-            vpmaxsd(vmm_index_y_itr, vmm_index_y_itr, vmm_zero);
-            // weight y1: shift weight_size
-            gather_i32_indices(vmm_weightY, reg_weight_y, sizeof(float), vmm_tbl_y, grid_len, Precision::FP32, true);
-            cubic_planar_line(true);
-
-            // y2
-            vpaddd(vmm_index_y_itr, vmm_index_in_y, vmm_one);
-            // crop to [0, IH - 1]
-            vpminsd(vmm_index_y_itr, vmm_index_y_itr, cubic_planar_table_val(1));
-            vpmaxsd(vmm_index_y_itr, vmm_index_y_itr, vmm_zero);
-            // weight y2
-            gather_i32_indices(vmm_weightY, reg_weight_y, 2 * sizeof(float), vmm_tbl_y, grid_len, Precision::FP32, true);
-            cubic_planar_line(true);
-
-            // y3
-            vpaddd(vmm_index_y_itr, vmm_index_in_y, vmm_one);
-            vpaddd(vmm_index_y_itr, vmm_index_y_itr, vmm_one);
-            // crop to [0, IH - 1]
-            vpminsd(vmm_index_y_itr, vmm_index_y_itr, cubic_planar_table_val(1));
-            vpmaxsd(vmm_index_y_itr, vmm_index_y_itr, vmm_zero);
-            // weight y3
-            gather_i32_indices(vmm_weightY, reg_weight_y, 3 * sizeof(float), vmm_tbl_y, grid_len, Precision::FP32, true);
-            cubic_planar_line(true);
-
-            if (attr_.post_ops_.len() != 0) {
-                apply_post_ops(jcp_.dst_prc, true);  // oc_off is broadcast and always the same value for this channel
-            }
-            store(vmm_val, reg_dst, scalar_step);
-
-            add(reg_tbl_y, scalar_step * sizeof(int));  // sizeof(int): sequence with dd()
-            add(reg_tbl_x, scalar_step * sizeof(int));
-            add(reg_dst, scalar_step * jcp_.dst_data_size);
-
-            sub(reg_work_amount, scalar_step);
-
-            jmp(tail_loop_label, T_NEAR);
-        }
-        L(tail_loop_end_label);
-    }
-
-    inline void cubic_planar_line(bool is_scalar) {
-        uni_vpxor(vmm_dstX, vmm_dstX, vmm_dstX);
-        cubic_planar_pixel(0, is_scalar);
-        cubic_planar_pixel(1, is_scalar);
-        cubic_planar_pixel(2, is_scalar);
-        cubic_planar_pixel(3, is_scalar);
-        uni_vfmadd231ps(vmm_val, vmm_dstX, vmm_weightY);
-    }
-
-    inline void cubic_planar_pixel(int itr, bool is_scalar) {
-        // vmm_index_in_x have index for src
-        if (itr == 0) {
-            vpsubd(vmm_index_x_itr, vmm_index_in_x, vmm_one);
-        } else if (itr == 1) {
-            vpaddd(vmm_index_x_itr, vmm_index_in_x, vmm_zero);
-        } else if (itr == 2) {
-            vpaddd(vmm_index_x_itr, vmm_index_in_x, vmm_one);
-        } else if (itr == 3) {
-            vpaddd(vmm_index_x_itr, vmm_index_in_x, vmm_one);
-            vpaddd(vmm_index_x_itr, vmm_index_x_itr, vmm_one);
-        }
-
-        // crop to [0, IW - 1]
-        vpminsd(vmm_index_x_itr, vmm_index_x_itr, cubic_planar_table_val(2));
-        vpmaxsd(vmm_index_x_itr, vmm_index_x_itr, vmm_zero);
-
-        // value
-        // index is: ptr[reg_src + (vmm_index_y_itr * jcp_.IW + vmm_index_x_itr) * jcp_.src_data_size]
-        uni_vmovdqu(vmm_mask, cubic_planar_table_val(2));
-        vpaddd(vmm_mask, vmm_mask, vmm_one);  // (IW - 1) + 1 = IW
-        uni_vpmulld(vmm_mask, vmm_mask, vmm_index_y_itr);
-        uni_vpaddd(vmm_index_x_itr, vmm_index_x_itr, vmm_mask);
-        gather_i32_indices(vmm_src, reg_src, 0, vmm_index_x_itr, jcp_.src_data_size, Precision::FP32, is_scalar);
-
-        if (itr == 0) {
-            uni_vfmadd231ps(vmm_dstX, vmm_src, vmm_weightX0);
-        } else if (itr == 1) {
-            uni_vfmadd231ps(vmm_dstX, vmm_src, vmm_weightX1);
-        } else if (itr == 2) {
-            uni_vfmadd231ps(vmm_dstX, vmm_src, vmm_weightX2);
-        } else if (itr == 3) {
-            uni_vfmadd231ps(vmm_dstX, vmm_src, vmm_weightX3);
-        }
-    }
-
-    inline void prepare_cubic_planar_table() {
-        auto broadcast_int = [&](int val) {
-            for (size_t d = 0; d < vlen / sizeof(int); ++d) {
-                dd(val);
-            }
-        };
-
-        align(64);
-        L(l_table_constant);
-        broadcast_int(vals_for_cubic_planar.int_one);
-        broadcast_int(jcp_.IH - 1);
-        broadcast_int(jcp_.IW - 1);
-        dd(vals_for_cubic_planar.mask_gather_avx512);
-    }
-
-    struct vals_for_cubic_planar_type {
-        int int_one = 0x00000001;
-        int mask_gather_avx512 = 0x0000ffff;  // 00000000000000001111111111111111
-    } vals_for_cubic_planar;
-
-    inline Xbyak::Address cubic_planar_table_val(int index) {
-        return ptr[reg_table + index * vlen];
-    }
-
-    // always gather to Vmm, compute with Vmm, store with Xmm if scalar_step
-    inline void gather_i32_indices(Vmm vmm_src, const Xbyak::Reg64 &base, int offset, Vmm vmm_indices, int scale,
-                                Precision src_prc, bool is_scalar) {
-        Xbyak::Address table_idx = ptr[base + offset + vmm_indices * scale];
-        if ((isa == cpu::x64::avx512_core) && !is_scalar) {
-            // [0-15] bit of int to mask
-            kmovw(k_mask, cubic_planar_table_val(3));
-            if (src_prc == Precision::FP32) {
-                vgatherdps(vmm_src | k_mask, table_idx);  // dword index, packed single data
-            } else if (src_prc == Precision::I32) {
-                vpgatherdd(vmm_src | k_mask, table_idx);  // dword index, dword data
-            }
-        } else if ((isa == cpu::x64::avx2) && !is_scalar) {
-            uni_vpcmpeqd(vmm_mask, vmm_mask, vmm_mask);
-            if (src_prc == Precision::FP32) {
-                vgatherdps(vmm_src, table_idx, vmm_mask);
-            } else if (src_prc == Precision::I32) {
-                vpgatherdd(vmm_src, table_idx, vmm_mask);
-            }
-        } else {
-            const int gpr_size = 8;
-            sub(rsp, gpr_size);
-            // move content in register to content in address(ptr[])
-            mov(ptr[rsp], reg_tmp_64);
-
-            // replace index with value in stack
-            sub(rsp, vlen);
-            uni_vmovdqu(ptr[rsp], vmm_indices);
-
-            int repeats = is_scalar ? 1 : vlen / sizeof(float);
-            for (size_t i = 0; i < repeats; ++i) {
-                mov(reg_tmp_64.cvt32(), ptr[rsp + i * sizeof(int)]);       // sizeof(int)  index_size
-                table_idx = ptr[base + offset + reg_tmp_64 * scale];       // scale: sizeof(float)   value_size
-                mov(reg_tmp_64.cvt32(), table_idx);
-                mov(ptr[rsp + i * sizeof(int)], reg_tmp_64.cvt32());
-            }
-
-            uni_vmovups(vmm_src, ptr[rsp]);
-            add(rsp, vlen);
-            // restore GPR state
-            mov(reg_tmp_64, ptr[rsp]);
-            add(rsp, gpr_size);
-        }
-    }
-
-    // is_broadcast for broadcasting param for depth_wise and quantize(channel-sensitive post-ops), for fusion with plain layout.
-    void apply_post_ops(Precision dst_prc, bool is_broadcast) {
-        const auto &p = attr_.post_ops_;
-        int eltwise_inj_idx = 0;
-        int depthwise_inj_idx = 0;
-        int quantization_inj_idx = 0;
-        int post_ops_data_offset = 0;
-        for (int i = 0; i < p.len(); i++) {
-            auto& post_op = p.entry_[i];
-            if (post_op.is_eltwise()) {
-                eltwise_injectors[eltwise_inj_idx]->compute_vector_range(vmm_val.getIdx(), vmm_val.getIdx() + 1);
-                eltwise_inj_idx++;
-            } else if (post_op.is_depthwise()) {
-                mov(reg_d_weights, ptr[reg_post_ops_data + post_ops_data_offset]);
-                add(reg_d_weights, reg_oc_off);
-
-                // weight and bias is padded. scalar as vector.
-                depthwise_injectors[depthwise_inj_idx]->compute_vector_range(
-                        vmm_val.getIdx(), vmm_val.getIdx() + 1, reg_d_weights, reg_d_weights, is_broadcast);
-
-                depthwise_inj_idx++;
-                post_ops_data_offset += depthwise_injectors[depthwise_inj_idx]->memoryStep();
-            } else if (post_op.is_quantization()) {
-                bool do_dequantization = post_op.quantization.alg == alg_kind::quantization_quantize_dequantize;
-                bool do_rounding = do_dequantization || dst_prc == Precision::FP32 || i != p.len() - 1;
-
-                int s_idx = vmm_val.getIdx();
-
-                quantization_injectors[quantization_inj_idx]->init_crop_ptrs(reg_post_ops_data + post_ops_data_offset, reg_oc_off);
-                quantization_injectors[quantization_inj_idx]->compute_crop(s_idx, s_idx + 1, 0, 0, is_broadcast);
-
-                quantization_injectors[quantization_inj_idx]->init_input_scale_shift_ptrs(reg_post_ops_data + post_ops_data_offset, reg_oc_off);
-                quantization_injectors[quantization_inj_idx]->compute_input_scale_shift(s_idx, s_idx + 1, 0, do_rounding, 0, is_broadcast);
-
-                if (do_dequantization) {
-                    quantization_injectors[quantization_inj_idx]->init_output_scale_shift_ptrs(reg_post_ops_data + post_ops_data_offset, reg_oc_off);
-                    quantization_injectors[quantization_inj_idx]->compute_output_scale_shift(s_idx, s_idx + 1, 0, 0, is_broadcast);
-                }
-
-                post_ops_data_offset += quantization_injectors[quantization_inj_idx]->memoryStep();
-                quantization_inj_idx++;
-            }
-        }
-    }
-};
 
 namespace {
 struct InterpolateKey {
     InterpolateAttrs nodeAttrs;
     VectorDims srcDims;
     VectorDims dstDims;
-    std::vector<float> dataScales;
     dnnl::primitive_attr attr;
 
     size_t hash() const;
@@ -1399,7 +63,7 @@ size_t InterpolateKey::hash() const {
 
     seed = get_vector_hash(seed, srcDims);
     seed = get_vector_hash(seed, dstDims);
-    seed = get_vector_hash(seed, dataScales);
+    seed = get_vector_hash(seed, nodeAttrs.dataScales);
 
     seed = hash_combine(seed, get_attr_hash(*attr.get()));
     return seed;
@@ -1433,7 +97,7 @@ bool InterpolateKey::operator==(const InterpolateKey &rhs) const {
         return false;
     if (dstDims != rhs.dstDims)
         return false;
-    if (dataScales != rhs.dataScales)
+    if (nodeAttrs.dataScales != rhs.nodeAttrs.dataScales)
         return false;
     if (!(*attr.get() == *rhs.attr.get()))
         return false;
@@ -1669,6 +333,8 @@ Interpolate::Interpolate(const std::shared_ptr<ngraph::Node>& op, const GraphCon
 }
 
 void Interpolate::getSupportedDescriptors() {
+    std::cout << "getSupportedDescriptors" << std::endl;
+
     if (getParentEdges().size() != 3 && getParentEdges().size() != 4)
         // data, target_shape, scale, axis(optional).
         IE_THROW() << errorPrefix << " has incorrect number of input edges";
@@ -1756,7 +422,22 @@ void Interpolate::initSupportedPrimitiveDescriptors() {
             config.inConfs[AXES_ID].setMemDesc(creatorsMap.at(LayoutType::ncsp)->createSharedDesc(axesType, getInputShapeAtPort(AXES_ID)));
 
         config.outConfs[0].setMemDesc(creatorsMap.at(dataFormat)->createSharedDesc(outputPrecision, getOutputShapeAtPort(0)));
-        supportedPrimitiveDescriptors.push_back({config, implDetail});
+#if defined(OPENVINO_ARCH_X86_64)
+        supportedPrimitiveDescriptors.push_back({config, impl_type});
+#else
+        std::vector<MemoryDescPtr> srcMemoryDescs;
+        for (int i = 0; i < config.inConfs.size(); i++) {
+            srcMemoryDescs.push_back(config.inConfs[i].getMemDesc());
+        }
+        std::vector<MemoryDescPtr> dstMemoryDescs;
+        for (int i = 0; i < config.outConfs.size(); i++) {
+            dstMemoryDescs.push_back(config.outConfs[i].getMemDesc());
+        }
+
+        auto factory = std::make_shared<InterpolateExecutorFactory>(interpAttrs, srcMemoryDescs, dstMemoryDescs,
+                                                                    std::make_shared<ExecutorContext>(context, getPrimitivesPriority()));
+        supportedPrimitiveDescriptors.push_back({config, implDetail, factory});
+#endif
     };
 
     const auto &dataMinDims = getInputShapeAtPort(DATA_ID).getMinDims();
@@ -1835,7 +516,7 @@ bool Interpolate::needPrepareParams() const {
     return (inputShapesModified() || lastOutputDims != getChildEdgesAtPort(0)[0]->getMemory().getStaticDims());
 }
 
-void Interpolate::prepareParams() {
+void Interpolate::prepareParams() {\
     if (!shapesDefined()) {
         IE_THROW() << "Can't prepare params for Interpolate node with name: " << getName() << ", because input/output dims aren't defined";
     }
@@ -1870,14 +551,13 @@ void Interpolate::prepareParams() {
         scales.assign(scalesData, scalesData + scalesMem.getStaticDims()[0]);
     }
 
-    std::vector<float> dataScales = getScales(getPaddedInputShape(srcDims, interpAttrs.padBegin, interpAttrs.padEnd), dstDims);
+    auto dataScales = getScales(getPaddedInputShape(srcDims, interpAttrs.padBegin, interpAttrs.padEnd), dstDims);
     if (getOutputShapeAtPort(0).getRank() > 2 && (dataScales[0] != 1.f || dataScales[1] != 1.f)) {
         IE_THROW() << "Interpolate layer only supports resize on spatial dimensions(depth, height and width)";
     }
-
-    InterpolateKey key = {interpAttrs, srcDims, dstDims, dataScales, dnnl::primitive_attr()};
+    InterpolateKey key = {interpAttrs, srcDims, dstDims, dnnl::primitive_attr()};
     setPostOps(key.attr, dstDims);
-
+#if defined(OPENVINO_ARCH_X86_64)
     auto buildExecutor = [&](const InterpolateKey& key) -> std::shared_ptr<InterpolateExecutor> {
         std::shared_ptr<InterpolateExecutor> executor;
         if ((key.nodeAttrs.mode == InterpolateMode::nearest || key.nodeAttrs.mode == InterpolateMode::linear_onnx ||
@@ -1887,13 +567,8 @@ void Interpolate::prepareParams() {
             executor = std::make_shared<InterpolateJitExecutor>(key.nodeAttrs,
                                                                key.srcDims,
                                                                key.dstDims,
-                                                               key.dataScales,
+                                                               key.nodeAttrs.dataScales,
                                                                key.attr);
-        } else {
-            executor = std::make_shared<InterpolateRefExecutor>(key.nodeAttrs,
-                                                               key.srcDims,
-                                                               key.dstDims,
-                                                               key.dataScales);
         }
         return executor;
     };
@@ -1901,7 +576,19 @@ void Interpolate::prepareParams() {
     auto cache = context->getParamsCache();
     auto result = cache->getOrCreate(key, buildExecutor);
     execPtr = result.first;
-
+#else
+    std::vector<MemoryDescPtr> srcMemoryDescs;
+    for (int i = 0; i < getOriginalInputsNumber(); i++) {
+        srcMemoryDescs.push_back(getParentEdgeAt(i)->getMemoryPtr()->getDescPtr());
+    }
+    std::vector<MemoryDescPtr> dstMemoryDescs;
+    for (int i = 0; i < getOriginalOutputsNumber(); i++) {
+        dstMemoryDescs.push_back(getChildEdgeAt(i)->getMemoryPtr()->getDescPtr());
+    }
+    auto selectedPD = getSelectedPrimitiveDescriptor();
+    execPtr = selectedPD->getExecutorFactoryAs<InterpolateExecutorFactory>()->makeExecutor(interpAttrs, srcMemoryDescs, dstMemoryDescs, key.attr);
+    selectedPD->setImplementationType(execPtr->getImplType());
+#endif
     lastOutputDims = dstDims;
 }
 
@@ -1979,7 +666,6 @@ void Interpolate::execute(dnnl::stream strm) {
     if (!execPtr) {
         IE_THROW() << "Can't execute Interpolate node. Primitive didn't created";
     }
-
     auto &dstMemPtr = getChildEdgeAt(0)->getMemoryPtr();
     auto &srcMemPtr = getParentEdgeAt(DATA_ID)->getMemoryPtr();
 
@@ -2059,355 +745,6 @@ void Interpolate::execute(dnnl::stream strm) {
     }
 
     execPtr->exec(src_data, dst_data, postOpsDataPtrs.data());
-}
-
-// for ndhwc and nCdhw8c[16c]
-// input may be f32/bf16/int8, fused->output varies
-void Interpolate::InterpolateJitExecutor::NNCGathered(const uint8_t *in_ptr_, uint8_t *out_ptr_, const void *post_ops_data_,
-                                                                int B, int C, int ID, int IH, int IW, int OD, int OH, int OW) {
-    int *index_d = static_cast<int*>(&indexTable[0]);
-    int *index_h = static_cast<int*>(&indexTable[OD]);
-    int *index_w = static_cast<int*>(&indexTable[OD + OH]);
-
-    bool is_nhwc = (configured_for_layout == by_channel);
-
-    for (int b = 0; b < B; b++) {
-        if (is_nhwc) {
-            const uint8_t *in_ptr = in_ptr_ + (IW * IH * ID * C * b) * srcDataSize;
-            uint8_t *out_ptr = out_ptr_ + (OW * OH * OD * C * b) * dstDataSize;
-            std::vector<int> index_w_kernel(OW);
-            for (int ox = 0; ox < OW; ox++) {
-                index_w_kernel[ox] = index_w[ox] * C * srcDataSize;
-            }
-            parallel_for2d(OD, OH, [&](size_t d, size_t h) {
-                // kernel for C * OW
-                uint8_t *out_ptr_dh = out_ptr + (C * OW * OH * d + C * OW * h) * dstDataSize;
-                const uint8_t *in_ptr_dh = in_ptr + (C * IW * IH * index_d[d] + C * IW * index_h[h]) * srcDataSize;
-                auto arg = jit_interpolate_call_args();
-                arg.dst = out_ptr_dh;
-                arg.src_ptr[0] = in_ptr_dh;
-                arg.index = static_cast<int*>(&(index_w_kernel[0]));
-                arg.work_amount = C;
-                arg.oc_off = 0;
-                arg.post_op_data = post_ops_data_;
-                (*interpolateKernel)(&arg);
-            });
-        } else {  // for blk
-            int blk_size = mayiuse(cpu::x64::avx512_core) ? 16 : 8;
-            int CB = div_up(C, blk_size);
-            const uint8_t *in_ptr = in_ptr_ + (IW * IH * ID * CB * blk_size * b) * srcDataSize;
-            uint8_t *out_ptr = out_ptr_ + (OW * OH * OD * CB * blk_size * b) * dstDataSize;
-            std::vector<int> index_w_kernel(OW);
-            for (int ox = 0; ox < OW; ox++) {
-                index_w_kernel[ox] = index_w[ox] * blk_size * srcDataSize;
-            }
-            parallel_for2d(CB, OD, [&](size_t cb, size_t d) {
-                uint8_t *out_ptr_cbd = out_ptr + (blk_size * OW * OH * OD * cb + blk_size * OW * OH * d) * dstDataSize;
-                const uint8_t *in_ptr_cbd = in_ptr + (blk_size * IW * IH * ID * cb + blk_size * IW * IH * index_d[d]) * srcDataSize;
-                auto arg = jit_interpolate_call_args();
-                for (int h = 0; h < OH; h++) {  // kernel for blk_size * OW
-                    arg.dst = out_ptr_cbd + blk_size * OW * h * dstDataSize;
-                    arg.src_ptr[0] = in_ptr_cbd + blk_size * IW * index_h[h] * srcDataSize;
-                    arg.index = static_cast<int*>(&(index_w_kernel[0]));
-                    arg.work_amount = static_cast<size_t>(OW);
-                    arg.oc_off = cb * blk_size * sizeof(float);
-                    arg.post_op_data = post_ops_data_;
-                    (*interpolateKernel)(&arg);
-                }
-            });
-        }
-    }  // batch end
-}
-
-void Interpolate::InterpolateJitExecutor::NNPlanar(const uint8_t *in_ptr_, uint8_t *out_ptr_, const void *post_ops_data_,
-                                                             int B, int C, int ID, int IH, int IW, int OD, int OH, int OW) {
-    int *index_d = static_cast<int*>(&indexTable[0]);
-    int *index_h = static_cast<int*>(&indexTable[OD]);
-    int *index_w = static_cast<int*>(&indexTable[OD + OH]);
-
-    std::vector<int> index_kernel(OH + OW);
-    // index_h * IW * srcDataSize to reduce and simplify redundant compute
-    for (int oh = 0; oh < OH; oh++) {
-        index_kernel[oh] = index_h[oh] * IW * srcDataSize;
-    }
-    // index_w * srcDataSize
-    for (int ow = 0; ow < OW; ow++) {
-        index_kernel[OH + ow] = index_w[ow] * srcDataSize;
-    }
-
-    parallel_for3d(B, C, OD, [&](size_t b, size_t c, size_t od) {
-        const uint8_t *in_ptr = in_ptr_ + (IW * IH * ID * C * b + IW * IH * ID * c + IW * IH * index_d[od]) * srcDataSize;
-        uint8_t *out_ptr = out_ptr_ + (OW * OH * OD * C * b + OW * OH * OD * c + OW * OH * od) * dstDataSize;
-
-        auto arg = jit_interpolate_call_args();
-        arg.src_ptr[0] = in_ptr;
-        arg.dst = out_ptr;
-        arg.index = static_cast<int*>(&index_kernel[0]);  // need index_h and index_w in kernel, it's in continous memory so one param
-        arg.oc_off = static_cast<size_t>(c * sizeof(float));
-        // work_amount is OH(out loop) and OW(inner loop), can get in kernel from jcp.
-        arg.post_op_data = post_ops_data_;
-        (*interpolateKernel)(&arg);
-    });
-}
-
-void Interpolate::InterpolateJitExecutor::linearOnnxPlanar(const uint8_t *in_ptr_, uint8_t *out_ptr_, const void *post_ops_data_, int B, int C,
-                                                                     int ID, int IH, int IW, int OD, int OH, int OW) {
-    // FrontTopLeft:0, FrontTopRight:1, FrontBottomLeft:2, FrontBottomRight:3, EndTopLeft:4,   EndTopRight:5,   EndBottomLeft:6,   EndBottomRight:7
-    // weight: Left:0, ritht:1, top:2, bottom:3, front:4, end:5
-    int *index = static_cast<int*>(&indexTable[0]);
-    int eltInGrid = (spatialDimSize > 2) ? MAX_INPUT_INTERPOLATE : ((spatialDimSize > 1) ? 4 : 2);
-    int scratchLen = rnd_up(eltInGrid * OW * OH * OD, 16);
-    float *weight = reinterpret_cast<float*>(&indexTable[scratchLen]);
-
-    parallel_for2d(B, C, [&](size_t b, size_t c) {
-        uint8_t *out_ptr_nc = out_ptr_ + (OH * OW * OD * C * b + OH * OW * OD * c) * dstDataSize;
-        const uint8_t *in_ptr_nc = in_ptr_ + (IH * IW * ID * C * b + IH * IW * ID * c) * srcDataSize;
-        auto arg = jit_interpolate_call_args();
-        arg.src_ptr[0] = in_ptr_nc;
-        arg.index = static_cast<int*>(&index[0]);
-        arg.weight_ptr[0] = static_cast<float*>(&weight[0]);
-        arg.dst = out_ptr_nc;
-        arg.work_amount = OW * OH * OD;
-        arg.oc_off = static_cast<size_t>(c * sizeof(float));
-        arg.post_op_data = post_ops_data_;
-        (*interpolateKernel)(&arg);
-    });
-}
-
-void Interpolate::InterpolateJitExecutor::linearOnnxCGathered(const uint8_t *in_ptr_, uint8_t *out_ptr_, const void *post_ops_data_,
-                                                                        int B, int C, int ID, int IH, int IW, int OD, int OH, int OW) {
-    // left:OW right:OW Top:OH Bottom:OH Front:OD End:OD
-    std::vector<int*> indexPtr(MAX_INPUT_INTERPOLATE, 0);
-    std::vector<float*> weightPtr(MAX_INPUT_INTERPOLATE, 0);
-    size_t scratchLen = rnd_up(OW + OW + OH + OH + OD + OD, 16);
-    indexPtr[0] = static_cast<int*>(&indexTable[0]);
-    indexPtr[1] = static_cast<int*>(&indexTable[OW]);
-    indexPtr[2] = static_cast<int*>(&indexTable[2 * OW]);
-    indexPtr[3] = static_cast<int*>(&indexTable[2 * OW + OH]);
-    indexPtr[4] = static_cast<int*>(&indexTable[2 * OW + 2 * OH]);
-    indexPtr[5] = static_cast<int*>(&indexTable[2 * OW + 2 * OH + OD]);
-
-    weightPtr[0] = reinterpret_cast<float*>(&indexTable[scratchLen]);
-    weightPtr[1] = reinterpret_cast<float*>(&indexTable[scratchLen + OW]);
-    weightPtr[2] = reinterpret_cast<float*>(&indexTable[scratchLen + 2 * OW]);
-    weightPtr[3] = reinterpret_cast<float*>(&indexTable[scratchLen + 2 * OW + OH]);
-    weightPtr[4] = reinterpret_cast<float*>(&indexTable[scratchLen + 2 * OW + 2 * OH]);
-    weightPtr[5] = reinterpret_cast<float*>(&indexTable[scratchLen + 2 * OW + 2 * OH + OD]);
-
-    bool isByChannel = (configured_for_layout == by_channel) ? true : false;
-
-    int blkSize = mayiuse(cpu::x64::avx512_core) ? 16 : 8;
-    int CB = isByChannel ? 1 : div_up(C, blkSize);
-    int CGatherLen = isByChannel ? C : blkSize;
-    int workAmount = isByChannel ? C : CB;
-    // n_CB(1)_d_h_w_8[16](c), () for by-channel
-    int C0 = OW * CGatherLen;
-    int C1 = OH * C0;
-    int C2 = OD * C1;
-    int C3 = CB * C2;
-    int I0 = IW * CGatherLen;
-    int I1 = IH * I0;
-    int I2 = ID * I1;
-    int I3 = CB * I2;
-    parallel_for3d(B, OD, OH, [&](size_t b, size_t d, size_t h) {
-        uint8_t *out_ptr_ndh = out_ptr_ + (C3 * b + C1 * d + C0 * h) * dstDataSize;
-
-        const uint8_t *in_ptr_n = in_ptr_ + (I3 * b) * srcDataSize;
-        const uint8_t *in_ptr_nf = in_ptr_n + (indexPtr[4][d] * I1) * srcDataSize;
-        const uint8_t *in_ptr_nft = in_ptr_nf + (indexPtr[2][h] * I0) * srcDataSize;
-        const uint8_t *in_ptr_nfb = in_ptr_nf + (indexPtr[3][h] * I0) * srcDataSize;
-        const uint8_t *in_ptr_ne = in_ptr_n + (indexPtr[5][d] * I1) * srcDataSize;
-        const uint8_t *in_ptr_net = in_ptr_ne + (indexPtr[2][h] * I0) * srcDataSize;
-        const uint8_t *in_ptr_neb = in_ptr_ne + (indexPtr[3][h] * I0) * srcDataSize;
-        auto arg = jit_interpolate_call_args();
-        for (int w = 0; w < OW; ++w) {
-            uint8_t *out_ptr_ndhw = out_ptr_ndh + CGatherLen * w * dstDataSize;
-
-            arg.src_ptr[0] = in_ptr_nft + (indexPtr[0][w] * CGatherLen) * srcDataSize;
-            arg.src_ptr[1] = in_ptr_nft + (indexPtr[1][w] * CGatherLen) * srcDataSize;
-            arg.src_ptr[2] = in_ptr_nfb + (indexPtr[0][w] * CGatherLen) * srcDataSize;
-            arg.src_ptr[3] = in_ptr_nfb + (indexPtr[1][w] * CGatherLen) * srcDataSize;
-            arg.src_ptr[4] = in_ptr_net + (indexPtr[0][w] * CGatherLen) * srcDataSize;
-            arg.src_ptr[5] = in_ptr_net + (indexPtr[1][w] * CGatherLen) * srcDataSize;
-            arg.src_ptr[6] = in_ptr_neb + (indexPtr[0][w] * CGatherLen) * srcDataSize;
-            arg.src_ptr[7] = in_ptr_neb + (indexPtr[1][w] * CGatherLen) * srcDataSize;
-            arg.weight_ptr[0] = static_cast<float*>(&weightPtr[0][w]);
-            arg.weight_ptr[1] = static_cast<float*>(&weightPtr[1][w]);
-            arg.weight_ptr[2] = static_cast<float*>(&weightPtr[2][h]);
-            arg.weight_ptr[3] = static_cast<float*>(&weightPtr[3][h]);
-            arg.weight_ptr[4] = static_cast<float*>(&weightPtr[4][d]);
-            arg.weight_ptr[5] = static_cast<float*>(&weightPtr[5][d]);
-            arg.dst = out_ptr_ndhw;
-            arg.work_amount = workAmount;
-            arg.oc_off = 0;
-            arg.post_op_data = post_ops_data_;
-            (*interpolateKernel)(&arg);
-        }
-    });
-}
-
-void Interpolate::InterpolateJitExecutor::cubicCGathered(const uint8_t *in_ptr_, uint8_t *out_ptr_, const void *post_ops_data_,
-                                                                   int B, int C, int IH, int IW, int OH, int OW) {
-    const int idxNum = 1;
-    int *xOrigin = static_cast<int*>(&indexTable[0]);
-    float *xFactor = reinterpret_cast<float*>(&indexTable[OW]);
-    int *yOrigin = static_cast<int*>(&indexTable[(CUBIC_GRID_LEN + idxNum) * OW]);
-    float *yFactor = reinterpret_cast<float*>(&indexTable[(CUBIC_GRID_LEN + idxNum) * OW + OH]);
-
-    int blkSize = mayiuse(cpu::x64::avx512_core) ? 16 : 8;
-    int CB = div_up(C, blkSize);
-    int CSize = configured_for_layout == InterpolateLayoutType::by_channel ? C : blkSize * CB;
-    int CGatherLen = configured_for_layout == InterpolateLayoutType::by_channel ? C : blkSize;
-    int workAmount = configured_for_layout == InterpolateLayoutType::by_channel ? C : CB;
-
-    parallel_for3d(B, OH, OW, [&](size_t b, size_t h, size_t w) {
-        uint8_t *out_ptr_nhw = out_ptr_ + (OH * OW * CSize * b + OW * CGatherLen * h + CGatherLen * w) * dstDataSize;
-        const uint8_t *in_ptr_n = in_ptr_ + (IH * IW * CSize * b) * srcDataSize;
-
-        std::vector<int> kernelIndex(CUBIC_GRID_LEN * CUBIC_GRID_LEN);  // 16 address offset to src(batch) or src(CB)
-        int iy = yOrigin[h];
-        int ix = xOrigin[w];
-        for (int y = iy - 1, i = 0; y <= iy + 2; y++, i++) {
-            int yInRange = std::max(0, std::min(y, IH - 1));
-            yInRange = yInRange * CGatherLen * IW * srcDataSize;
-            for (int x = ix - 1, j = 0; x <= ix + 2; x++, j++) {
-                int xInRange = std::max(0, std::min(x, IW - 1));
-                xInRange = yInRange + xInRange * CGatherLen * srcDataSize;
-                kernelIndex[i * CUBIC_GRID_LEN + j] = xInRange;
-            }
-        }
-        auto arg = jit_interpolate_call_args();
-            arg.dst = out_ptr_nhw;
-            arg.src_ptr[0] = in_ptr_n;
-            arg.index = static_cast<int*>(&kernelIndex[0]);
-            // 0 for weight_W, 1 for weight_H
-            arg.weight_ptr[0] = static_cast<float*>(&xFactor[w * CUBIC_GRID_LEN]);
-            arg.weight_ptr[1] = static_cast<float*>(&yFactor[h * CUBIC_GRID_LEN]);
-
-            // for by channel, src + step, dst + step, process next step on continuous memory
-            // for blk, src + IW*IH*blkSize, dst + OW*OH*blkSize, process the blkSize on next CB
-            arg.work_amount = workAmount;
-            arg.oc_off = 0;
-            arg.post_op_data = post_ops_data_;
-            (*interpolateKernel)(&arg);
-    });
-}
-
-void Interpolate::InterpolateJitExecutor::cubicPlanar(const uint8_t *in_ptr_, uint8_t *out_ptr_, const void *post_ops_data_,
-                                                                int B, int C, int IH, int IW, int OH, int OW) {
-    int tblAdvance = 0;
-    int *xOrigin = static_cast<int*>(&indexTable[tblAdvance]);
-    tblAdvance += OW;
-    float *xFactor = reinterpret_cast<float*>(&indexTable[tblAdvance]);
-    tblAdvance += CUBIC_GRID_LEN * OW;
-    int *yOrigin = static_cast<int*>(&indexTable[tblAdvance]);
-    tblAdvance += OH;
-    float *yFactor = reinterpret_cast<float*>(&indexTable[tblAdvance]);
-
-    tblAdvance += CUBIC_GRID_LEN * OH;
-    int *sequenceOH = static_cast<int*>(&indexTable[tblAdvance]);
-    tblAdvance += OW * OH;
-    int *sequenceOW = static_cast<int*>(&indexTable[tblAdvance]);
-
-    parallel_for2d(B, C, [&](size_t n, size_t c) {
-        const uint8_t *in_ptr_nc = in_ptr_ + (IW * IH * C * n + IW * IH * c) * srcDataSize;
-        uint8_t *out_ptr_nc = out_ptr_ + (OW * OH * C * n + OW * OH * c) * dstDataSize;
-
-        auto arg = jit_interpolate_call_args();
-        arg.dst = out_ptr_nc;
-        arg.src_ptr[0] = in_ptr_nc;
-        arg.index = xOrigin;
-        arg.src_ptr[1] = yOrigin;
-        arg.src_ptr[2] = static_cast<int*>(&sequenceOH[0]);
-        arg.src_ptr[3] = static_cast<int*>(&sequenceOW[0]);
-        arg.weight_ptr[0] = xFactor;
-        arg.weight_ptr[1] = yFactor;
-        arg.work_amount = static_cast<size_t>(OW * OH);
-        arg.oc_off = static_cast<size_t>(c * sizeof(float));
-        arg.post_op_data = post_ops_data_;
-        (*interpolateKernel)(&arg);
-    });
-}
-
-Interpolate::InterpolateJitExecutor::InterpolateJitExecutor(const InterpolateAttrs& interpAttrs,
-                                                                      const VectorDims &srcDims,
-                                                                      const VectorDims &dstDims,
-                                                                      const std::vector<float> &dataScales,
-                                                                      const dnnl::primitive_attr &attr) :
-        InterpolateExecutor(interpAttrs, srcDims, dstDims, dataScales) {
-    auto jcp = jit_interpolate_config_params();
-    jcp.mode = mode;
-    jcp.src_prc = interpAttrs.inPrc;
-    jcp.dst_prc = interpAttrs.outPrc;
-    jcp.src_data_size = jcp.src_prc.size();
-    jcp.dst_data_size = jcp.dst_prc.size();
-    jcp.indices_size = sizeof(int);
-    jcp.C = dstDim5d[1];
-    jcp.OW = dstDim5d[4];
-    jcp.OH = dstDim5d[3];
-    jcp.OD = dstDim5d[2];
-    jcp.IW = srcDimPad5d[4];
-    jcp.IH = srcDimPad5d[3];
-    jcp.ID = srcDimPad5d[2];
-    jcp.spatial_dim_size = getSpatialDimsNum(srcDims.size());
-    jcp.layout = interpAttrs.layout;
-    if (jcp.layout != InterpolateLayoutType::planar) {
-        if (mayiuse(cpu::x64::avx512_core)) {
-            interpolateKernel.reset(new jit_uni_interpolate_kernel_f32<cpu::x64::avx512_core>(jcp, *attr.get()));
-        } else if (mayiuse(cpu::x64::avx2)) {
-            interpolateKernel.reset(new jit_uni_interpolate_kernel_f32<cpu::x64::avx2>(jcp, *attr.get()));
-        } else if (mayiuse(cpu::x64::sse41)) {
-            interpolateKernel.reset(new jit_uni_interpolate_kernel_f32<cpu::x64::sse41>(jcp, *attr.get()));
-        }
-    } else if (mayiuse(cpu::x64::avx2) && interpAttrs.inPrc == InferenceEngine::Precision::FP32) {
-        // gather ISA(for planar JIT kernel) for avx2 and fp32
-        interpolateKernel.reset(new jit_uni_interpolate_kernel_f32<cpu::x64::avx2>(jcp, *attr.get()));
-    } else {
-        IE_THROW() << "Can't create InterpolateJitExecutor";
-    }
-    if (interpolateKernel) {
-        interpolateKernel->create_ker();
-    } else {
-        IE_THROW() << "Can't compile InterpolateJitExecutor";
-    }
-}
-
-void Interpolate::InterpolateJitExecutor::exec(const uint8_t *in_ptr_, uint8_t *out_ptr_, const void *post_ops_data_) {
-    size_t N = srcDimPad5d[0], C = srcDimPad5d[1], ID = srcDimPad5d[2], IH = srcDimPad5d[3], IW = srcDimPad5d[4];
-    size_t OD = dstDim5d[2], OH = dstDim5d[3], OW = dstDim5d[4];
-
-    if (!interpolateKernel) {
-        IE_THROW() << "Can't execute, kernel for Interpolate node is not compiled";
-    }
-    switch (mode) {
-        case InterpolateMode::nearest: {
-            if (configured_for_layout == InterpolateLayoutType::planar) {
-                NNPlanar(in_ptr_, out_ptr_, post_ops_data_, N, C, ID, IH, IW, OD, OH, OW);
-            } else {
-                NNCGathered(in_ptr_, out_ptr_, post_ops_data_, N, C, ID, IH, IW, OD, OH, OW);
-            }
-            break;
-        }
-        case InterpolateMode::linear_onnx: {
-            if (configured_for_layout == InterpolateLayoutType::planar) {
-                linearOnnxPlanar(in_ptr_, out_ptr_, post_ops_data_, N, C, ID, IH, IW, OD, OH, OW);
-            } else {
-                linearOnnxCGathered(in_ptr_, out_ptr_, post_ops_data_, N, C, ID, IH, IW, OD, OH, OW);
-            }
-            break;
-        }
-        case InterpolateMode::cubic: {
-            if (configured_for_layout == InterpolateLayoutType::planar) {
-                cubicPlanar(in_ptr_, out_ptr_, post_ops_data_, N, C, IH, IW, OH, OW);
-            } else {
-                cubicCGathered(in_ptr_, out_ptr_, post_ops_data_, N, C, IH, IW, OH, OW);
-            }
-            break;
-        }
-        default: {
-            IE_THROW() << "InterpolateJitExecutor has unsupported interpolate mode: " << mode;
-        }
-    }
 }
 
 bool Interpolate::canFuse(const NodePtr& node) const {
